refactor: 전체 보기 페이지의 컴포넌트 구성을 노선 관리, 구간 관리와 일치시킨다
