+++ src/components/OverviewList/index.jsx
@@ -1,17 +1,20 @@
 import React from "react";
-import PropTypes from "prop-types";
+import { useSelector } from "react-redux";
+import { selectOverviewList } from "../../pages/Overview/slice";
 import OverviewListItem from "../OverviewListItem";
 
-const OverviewList = ({ list }) => (
-  <ul className="space-y-6">
-    {list.map((item) => (
-      <OverviewListItem key={item.id} item={item} />
-    ))}
-  </ul>
-);
+const OverviewList = () => {
+  const list = useSelector(selectOverviewList);
 
-OverviewList.propTypes = {
-  list: PropTypes.arrayOf(OverviewListItem.propTypes.item).isRequired,
+  return (
+    list.length > 0 && (
+      <ul className="space-y-6">
+        {list.map((item) => (
+          <OverviewListItem key={item.id} item={item} />
+        ))}
+      </ul>
+    )
+  );
 };
 
 export default OverviewList;
